SBERDOMA-128 Review Fixes
